+++ src/medium/stringio.rs
@@ -27,9 +27,9 @@ impl StreamIn for StringIn {
         false
     }
     fn lines(&self) -> Box<dyn NextLine + '_> {
-        let a = self.0.inner.lock().unwrap().take().unwrap();
-        let b = a.lines();
-        Box::new(Lines { buf: b })
+        Box::new(Lines {
+            buf: self.lock_bufread().lines(),
+        })
     }
 }
 
@@ -143,15 +143,15 @@ const LINE_BUF_SIZE: usize = 1024;
 
 #[derive(Debug)]
 struct LockableStringIn {
-    inner: Mutex<Option<BufReader<RawStringIn>>>,
+    inner: Mutex<BufReader<RawStringIn>>,
 }
 impl LockableStringIn {
     pub fn with(a_string: String) -> Self {
         LockableStringIn {
-            inner: Mutex::new(Some(BufReader::with_capacity(
+            inner: Mutex::new(BufReader::with_capacity(
                 LINE_BUF_SIZE,
                 RawStringIn::new(a_string),
-            ))),
+            )),
         }
     }
     pub fn lock(&self) -> LockableStringInLock<'_> {
@@ -168,22 +168,22 @@ impl Default for LockableStringIn {
 
 #[derive(Debug)]
 struct LockableStringInLock<'a> {
-    inner: MutexGuard<'a, Option<BufReader<RawStringIn>>>,
+    inner: MutexGuard<'a, BufReader<RawStringIn>>,
 }
 impl Read for LockableStringInLock<'_> {
     #[inline(always)]
     fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        self.inner.as_mut().unwrap().read(buf)
+        self.inner.read(buf)
     }
 }
 impl BufRead for LockableStringInLock<'_> {
     #[inline(always)]
     fn fill_buf(&mut self) -> std::io::Result<&[u8]> {
-        self.inner.as_mut().unwrap().fill_buf()
+        self.inner.fill_buf()
     }
     #[inline(always)]
     fn consume(&mut self, amt: usize) {
-        self.inner.as_mut().unwrap().consume(amt)
+        self.inner.consume(amt)
     }
 }
 
@@ -230,16 +230,16 @@ impl Write for LockableStringOutLock<'_> {
     }
 }
 
-pub struct Lines {
-    buf: std::io::Lines<BufReader<RawStringIn>>,
+pub struct Lines<'a> {
+    buf: std::io::Lines<Box<dyn BufRead + 'a>>,
 }
-impl Iterator for Lines {
+impl<'a> Iterator for Lines<'a> {
     type Item = Result<String>;
     fn next(&mut self) -> Option<Result<String>> {
         self.buf.next()
     }
 }
-impl NextLine for Lines {}
+impl<'a> NextLine for Lines<'a> {}
 
 #[derive(Debug)]
 struct RawStringIn {